+++ packages/video_player_avplay/example/lib/main.dart
@@ -69,12 +69,19 @@ class _HlsRomoteVideo extends StatefulWidget {
 
 class _HlsRomoteVideoState extends State<_HlsRomoteVideo> {
   late VideoPlayerController _controller;
+  // Optional
+  final Map<StreamingPropertyType, String> _streamingProperties =
+      <StreamingPropertyType, String>{
+    StreamingPropertyType.adaptiveInfo: 'STARTBITRATE=460560',
+    StreamingPropertyType.setAbr: '0',
+  };
 
   @override
   void initState() {
     super.initState();
     _controller = VideoPlayerController.network(
       'https://test-streams.mux.dev/x36xhzz/x36xhzz.m3u8',
+      streamingProperty: _streamingProperties,
     );
 
     _controller.addListener(() {
@@ -84,7 +91,12 @@ class _HlsRomoteVideoState extends State<_HlsRomoteVideo> {
       setState(() {});
     });
     _controller.setLooping(true);
-    _controller.initialize().then((_) => setState(() {}));
+    _controller.initialize().then((_) async {
+      setState(() {});
+      // Optional
+      final List<Track> activateTracks = await _controller.getActiveTrackInfo();
+      print('[getActivateTrackInfo()] result: $activateTracks');
+    });
     _controller.play();
   }
 
@@ -173,6 +185,7 @@ class _DashRomoteVideoState extends State<_DashRomoteVideo> {
       _controller
           .getStreamingProperty(StreamingPropertyType.subtitleStreamInfo);
       _controller.getStreamingProperty(StreamingPropertyType.videoStreamInfo);
+
       _controller.getData(<DashPlayerProperty>{DashPlayerProperty.httpHeader});
     });
     _controller.play();